requirements.txt
aiohttp==3.6.2
asgiref==3.2.10
async-timeout==3.0.1
asyncpg==0.21.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
Cybernator==0.5.0
discord==1.0.1
discord.py==1.3.4
Django==2.2.16
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-steamauth==1.1.1
idna==2.9
multidict==4.7.6
#mysqlclient @ file:///C:/Users/doros/Downloads/mysqlclient-1.4.6-cp38-cp38-win32.whl
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
tabulate==0.8.7
urllib3==1.25.10
websockets==8.1
yarl==1.4.2
